+++ src/cdm_data_loaders/pipelines/cts_defaults.py
@@ -1,26 +1,30 @@
 """Common defaults for running pipelines on the KBase CTS."""
 
-from typing import Any, Self
+from typing import Any, Final, Self
 
 import dlt.common.configuration.accessors
 from frozendict import frozendict
 from pydantic import AliasChoices, Field, computed_field, field_validator, model_validator
-from pydantic_settings import BaseSettings, CliSuppress, SettingsConfigDict
+from pydantic_settings import CliSuppress, SettingsConfigDict
 
-INPUT_MOUNT = "/input_dir"
-OUTPUT_MOUNT = "/output_dir"
-
-VALID_DESTINATIONS = ["local_fs", "s3"]
-DEFAULT_BATCH_SIZE = 50
-DEFAULT_START_AT = 0
+from cdm_data_loaders.utils.batcher import MIN_START_AT
+from cdm_data_loaders.utils.cdm_logger import ARG_ALIASES as LOGGER_ARG_ALIASES
+from cdm_data_loaders.utils.cdm_logger import DEFAULT_LOG_CONFIG_FILE, LoggerSettings
 
 # TODO: frozendict can be moved to the stdlib implementation when py 3.15 is released.
+
+
+INPUT_MOUNT: Final[str] = "/input_dir"
+OUTPUT_MOUNT: Final[str] = "/output_dir"
+
+VALID_DESTINATIONS: list[str] = ["local_fs", "s3"]
 
 
 DEFAULT_CTS_SETTINGS = frozendict(
     {
         "dev_mode": False,
         "input_dir": INPUT_MOUNT,
+        "log_config_file": DEFAULT_LOG_CONFIG_FILE,
         # N.b. this gets replaced by destination.local_fs.bucket_url
         "output": "",
         "use_destination": "local_fs",
@@ -31,9 +35,11 @@ DEFAULT_CTS_SETTINGS = frozendict(
 DEFAULT_BATCH_FILE_SETTINGS = frozendict(
     {
         **DEFAULT_CTS_SETTINGS,
-        "start_at": DEFAULT_START_AT,
+        "start_at": MIN_START_AT,
     }
 )
+
+DEFAULT_PIPELINE_BATCH_SIZE: Final[int] = 50
 
 
 ARG_ALIASES = frozendict(
@@ -41,6 +47,7 @@ ARG_ALIASES = frozendict(
         "batch_size": ["-b", "--batch_size", "--batch-size"],
         "dev_mode": ["--dev_mode", "--dev-mode", "--dev"],
         "input_dir": ["-i", "--input_dir", "--input-dir"],
+        "log_config_file": LOGGER_ARG_ALIASES,
         "output": ["-o", "--output"],
         "start_at": ["-s", "--start_at", "--start-at"],
         "use_destination": ["-d", "--use_destination", "--use-destination"],
@@ -62,7 +69,7 @@ DEFAULT_SETTINGS_CONFIG_DICT = frozendict(
 )
 
 
-class CtsSettings(BaseSettings):
+class CtsSettings(LoggerSettings):
     """Configuration for running a basic import pipeline."""
 
     model_config = SettingsConfigDict(**DEFAULT_SETTINGS_CONFIG_DICT)