+++ tests/pipelines/test_all_the_bacteria.py
@@ -490,7 +490,7 @@ def test_osf_file_downloader_success(
     ],
 )
 @pytest.mark.parametrize("use_destination", VALID_DESTINATIONS)
-def test_osf_file_downloader_error_handling(  # noqa: PLR0913
+def test_osf_file_downloader_error_handling(
     atb_file_list: list[dict[str, Any]],
     expected_exceptions: list[str],
     expected_paths: dict[str, bool],