+++ tests/pipelines/test_cts_defaults.py
@@ -8,11 +8,11 @@ from pydantic_settings import CliApp
 
 from cdm_data_loaders.pipelines.cts_defaults import (
     ARG_ALIASES,
-    DEFAULT_START_AT,
     VALID_DESTINATIONS,
     BatchedFileInputSettings,
     CtsSettings,
 )
+from cdm_data_loaders.utils.batcher import MIN_START_AT
 from tests.pipelines.conftest import (
     DEFAULT_BATCH_FILE_SETTINGS_RECONCILED,
     DEFAULT_CTS_SETTINGS_RECONCILED,
@@ -98,7 +98,7 @@ def test_settings_all_settings_specified(
 ) -> None:
     """Ensure the CTS settings are set up correctly."""
     s = make_settings(settings_cls, dlt_config=dlt_config, **args)
-    check_settings(s, dict(expected))
+    check_settings(s, expected)
 
 
 @pytest.mark.parametrize("settings_cls", SETTINGS_CLASSES)
@@ -108,7 +108,7 @@ def test_cli_app_run_default_settings(settings_cls: type[CtsSettings], dlt_confi
     expected = (
         DEFAULT_CTS_SETTINGS_RECONCILED if settings_cls == CtsSettings else DEFAULT_BATCH_FILE_SETTINGS_RECONCILED
     )
-    check_settings(s, dict(expected))
+    check_settings(s, expected)
 
 
 @pytest.mark.parametrize("settings_cls", SETTINGS_CLASSES)
@@ -287,6 +287,7 @@ def test_settings_trailing_slash_stripped(
 @pytest.mark.parametrize("settings_cls", SETTINGS_CLASSES)
 @pytest.mark.parametrize("dev_mode", ARG_ALIASES["dev_mode"])
 @pytest.mark.parametrize("input_dir", ARG_ALIASES["input_dir"])
+@pytest.mark.parametrize("log_config_file", ARG_ALIASES["log_config_file"])
 @pytest.mark.parametrize("output", ARG_ALIASES["output"])
 @pytest.mark.parametrize("start_at", ARG_ALIASES["start_at"])
 @pytest.mark.parametrize("use_destination", ARG_ALIASES["use_destination"])
@@ -294,10 +295,11 @@ def test_settings_trailing_slash_stripped(
     "use_output_dir_for_pipeline_metadata",
     ARG_ALIASES["use_output_dir_for_pipeline_metadata"],
 )
-def test_cli_app_run_alt_settings(  # noqa: PLR0913
+def test_cli_app_run_alt_settings(
     settings_cls: type[CtsSettings],
     dev_mode: str,
     input_dir: str,
+    log_config_file: str,
     output: str,
     start_at: str,
     use_destination: str,
@@ -311,6 +313,8 @@ def test_cli_app_run_alt_settings(  # noqa: PLR0913
         TEST_BATCH_FILE_SETTINGS["dev_mode"],
         input_dir,
         TEST_BATCH_FILE_SETTINGS["input_dir"],
+        log_config_file,
+        TEST_BATCH_FILE_SETTINGS["log_config_file"],
         output,
         TEST_BATCH_FILE_SETTINGS["output"],
         use_destination,
@@ -403,7 +407,7 @@ def test_settings_generate_pipeline_raw_data_dirs(
         "output": DESTINATION_TO_OUTPUT[use_destination] if output == "" else OUTPUT_PATHS[output][OUT],
     }
     if settings_cls == BatchedFileInputSettings:
-        expected["start_at"] = DEFAULT_START_AT
+        expected["start_at"] = MIN_START_AT
 
     if (OUTPUT_PATHS[expected["output"]][S3] and use_destination == "local_fs") or (
         OUTPUT_PATHS[expected["output"]][S3] is False and use_destination == "s3"
@@ -463,7 +467,7 @@ def test_cli_app_run_generate_pipeline_raw_data_dirs(
         "output": DESTINATION_TO_OUTPUT[use_destination] if output == "" else OUTPUT_PATHS[output][OUT],
     }
     if settings_cls == BatchedFileInputSettings:
-        expected["start_at"] = DEFAULT_START_AT
+        expected["start_at"] = MIN_START_AT
 
     if (OUTPUT_PATHS[expected["output"]][S3] and use_destination == "local_fs") or (
         OUTPUT_PATHS[expected["output"]][S3] is False and use_destination == "s3"